pyproject.toml
[project]
name = "pow2core"
version = "0.1.0"
description = "Pow2Core is a core library for Pow2 mining calculations"
readme = "README.md"
authors = [
    { name = "poplab" }
]
requires-python = ">=3.12.0"
dependencies = [
    "numpy>=2.3.2",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.1",
    "pytest>=8.0.0",
]

[tool.ruff]
# Exclude .history directory from linting
exclude = [
    ".history",
    ".git",
    "__pycache__",
    ".venv",
    ".pytest_cache",
    "build",
    "dist",
    "tests",
]

# Set the maximum line length
line-length = 88

# Set the Python version
target-version = "py312"

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    # "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "N",    # pep8-naming
    "PL",   # pylint
    "PT",   # flake8-pytest-style
]

# Disable specific rules
ignore = [
    "E501",   # line-too-long (handled by formatter)
    "B008",   # do-not-use-function-calls-in-argument-defaults
    "PLR0913", # too-many-arguments
    "PLR0915", # too-many-statements
    "PLR2004", # magic-value-comparison
    "N806",   # non-lowercase-variable-in-function
]


[tool.ruff.lint.isort]
known-first-party = ["pow2core"]

[tool.ruff.lint.pylint]
max-args = 8
